Fix big perf regression
Before we called setState once but now we call it for every single element. Because we're not inside of a normal react event listener, setState are not batched and are doing a full re-render for each of them.

This uses ReactDOM.unstable_batchedUpdates in order to bring back the batching.

Note: this doesn't fix everything. Reloading the page causes 3 renders instead of one. Same for mouseup. But this fixes the most egregorious perf issue.