🐛 Fix emoji rendering
In the Collaboration dialog, there are 2 emojis in paragraphs. The 🔒 one was always correctly displayed on macOS because the system font defaults to Emoji Presentation mode for this emoji, whereas the ⚠️ one rendered gray-scale because the system font defaults to Text Presentation mode for this emoji.

It may differ a lot based on the systems, so the most consistent way to have Emoji Presentation seems to be `font-family: sans-serif`. Even Variation Selector-16 didn't work here.