fix: improve mobile user experience
On mobile and tablet devices, user experience suffers as the footer
controls are below the viewport height, becoming unreachable due to
scrolling being blocked by pinch-to-zoom.

Additionally, using the contextual menus on long touch causes the menu
items text to be selected.

This commit fixes both issues by improving the page css,
disabling user selection on touch as well as finxing the
html and body height to 100%.