Adds migration for pasted items, updates migration for draw elements, restores draw type
There's no longer a "draw" type in Excalidraw: in the restoreElement function, we're turning all "draw" elements to "line" elements. New shapes created with the pen tool have the "freedraw" type (and are typed as `ExcalidrawFreeDrawElement`). However, I'm putting the "draw" type back into the codebase to assist with the migration.