feat: support a system theme in `DarkModeToggle` component
Add a new prop named `withSystemTheme` to allow developers to use a third value representing the system theme. This is an OPT-IN feature and will not break any existing code.

Since have 3 values now, the icon is not displaying the next state anymore but rather the current state, to avoid any confusion for the user.