refactor: wrap exported SVG elements in a group
All the drawn elements of the exported SVGs are now wrapped in a group, this allows to apply filters with CSS rules and to condition them with media queries. The media query will be used to export SVGs supporting the system theme.

As a side-effect, this new way to apply CSS filters to SVG elements "improves" the rendering in macOS Quick Look. Before, a dark theme SVG containing an image was rendered with the embedded image having inverted colors. Now, the SVG is no longer rendered with a dark theme, but the image has proper colors.
However, everything works as expected in a browser.