feature: support SVG export with system theme
The `DarkModeToggle` component in the export dialog now supports a third theme representing the system one. When selected, the exported SVG will contain a media query to conditionally apply the dark theme when `prefers-color-scheme: dark` is matched. If the SVG viewer doesn't support CSS, the SVG will simply be rendered with the light theme.

The `exportWithDarkMode` attribute in `AppState` has been replaced by `exportTheme`, which supports all the value declared in the `THEME` constant.