Enable eraser stylus events, execute via eraser tool.
Implements a standard semantic, where the eraser tool activates in the
eraser orientation without switching the primary tool.

This is activated for for pens, like the windows surface pens,
which implement the the [standard eraser button state].

Implemented via a "tool stack" on the pointer action.
If the eraser button is used, then switch to eraser for the span of the
pointer drag event, then pop back to the previous tool on button up.

[standard eraser button state]: https://developer.mozilla.org/en-US/docs/Web/API/Pointer_events#determining_button_states

Related to:
https://github.com/excalidraw/excalidraw/issues/5281
https://github.com/zsviczian/obsidian-excalidraw-plugin/issues/965