fix: add fallbacks for all fonts
In the case that the WOFF2 fonts cannot be requested from the various font URLs
eg. https://excalidraw.com/Virgil.woff2, then provide fallbacks for Windows,
macOS and Linux.

This can happen if the SVG is loaded in an environment with a Content Security
Policy set up, like if an SVG file is included in a readme on GitHub:

https://github.com/excalidraw/excalidraw/issues/4855