fix: confirming `clear canvas` via enter confirmed the confirm dialog immediately
This was due to the same `keydown` event being caught by the confirm dialog which autofocuses the `confirm` button.

This fix executes the command in a new tick, and same for autofocusing the button, either of which would fix this specific case, but I'm also doing the autofocusing so it fixes potential other or future issues.

Note that it doesn't fix for when you keep holding the enter, as the browser will keep emitting new events after about a second.